[src/map] moved IdentityGraph methods into a separated file
